# Conflicts:
#	src/main/java/com/example/jobpuzzle/user/entity/User.java

+++ src/main/java/com/example/jobpuzzle/report/entity/FinalReport.java
@@ -18,7 +18,7 @@ public class FinalReport {
 
     private Long sessionId;
 
-    private Double overallScore;
+    private Integer overallScore;
 
     private String scoreLabel;
